+++ src/main/java/gradation/implementation/presentationtier/form/LevelForm.java
@@ -5,7 +5,7 @@ public class LevelForm {
 
     private String name;
     private Byte place;
-    private Short maximumThreshold;
+    private Integer maximumThreshold;
     private Double ratioPoints;
 
     public String getName() {
@@ -24,11 +24,11 @@ public class LevelForm {
         this.place = place;
     }
 
-    public Short getMaximumThreshold() {
+    public Integer getMaximumThreshold() {
         return maximumThreshold;
     }
 
-    public void setMaximumThreshold(Short maximumThreshold) {
+    public void setMaximumThreshold(Integer maximumThreshold) {
         this.maximumThreshold = maximumThreshold;
     }
